feat: changing card color on Click
