+++ src/components/About.jsx
@@ -3,12 +3,12 @@ import Tilt from "react-parallax-tilt";
 import { motion } from "framer-motion";
 
 import { styles } from "../styles";
-import { cardTitles } from "../constants";
+import { cardTitles, titleIntro, presentationText } from "../constants";
 import { fadeIn, textVariant } from "../utils/motion";
 
 import { SectionWrapper } from "../highOrderComponents";
 
-const AboutCard = ({ index, title, icon, onClick }) => {
+const AboutCard = ({ index, title, icon, onClick, selected }) => {
 	return (
 		<Tilt className="xs:w-[250px] w-full">
 			<motion.div
@@ -18,11 +18,15 @@ const AboutCard = ({ index, title, icon, onClick }) => {
 			>
 				<div
 					options={{ max: 45, scale: 1, speed: 450 }}
-					className="bg-blue-brabo rounded-[20px] py-5 
-          px-12 min-h-[280px] flex justify-evenly items-center flex-col"
+					className={` rounded-[20px] py-5 px-12 min-h-[280px] flex justify-evenly items-center flex-col 
+					${selected ? "bg-yellow-200" : "bg-blue-brabo"}`}
 				>
 					<img src={icon} alt={title} className="w-16 h-16 object-contain" />
-					<h3 className="text-white text-[20px] font-bold text-center">
+					<h3
+						className={`text-[20px] font-bold text-center ${
+							selected ? "text-blue-brabo" : "text-white"
+						}`}
+					>
 						{title}
 					</h3>
 				</div>
@@ -33,58 +37,6 @@ const AboutCard = ({ index, title, icon, onClick }) => {
 
 const About = () => {
 	const [selectedCard, setSelectedCard] = useState(0);
-
-	const titleIntro = [
-		"Overview.",
-		"Backend Developer.",
-		"Energy Engineer.",
-		"Green Transition Enthusiast.",
-		"Cloud Computing and Devops.",
-	];
-
-	const presentationText = [
-		"A passionate Backend Developer and Energy Engineer, currently pursuing \
-an MSc in Software Design at ITU. With a keen interest in exploring the \
-intricacies of new technologies, I immerse myself in the vibrant world \
-of software and hardware development. My aspiration is to uncover the \
-synergy between technology and sustainability, discovering how these two \
-forces can work hand in hand.",
-
-		"As a dedicated Backend Developer, I thrive on \
-crafting the unseen backbone of applications, ensuring seamless data \
-management and optimal performance. My passion lies in designing robust \
-server-side architectures, meticulously handling databases, and building \
-efficient APIs. Embracing the ever-evolving landscape of programming \
-languages and frameworks, I am committed to delivering high-quality, \
-scalable, and secure solutions that empower the frontend to shine.",
-
-		"As an Energy Engineer, I am driven by the goal of \
-promoting sustainable energy practices and reducing the carbon footprint \
-of modern technology. My expertise lies in designing energy-efficient \
-systems, optimizing energy usage, and exploring renewable energy \
-sources. I believe in the transformative power of eco-conscious \
-engineering, and I am dedicated to shaping a greener future through \
-innovation and conscious decision-making.",
-
-		"As a fervent Green Transition Enthusiast, I am deeply committed to \
-advocating for environmental sustainability in every aspect of life. \
-From personal habits to professional endeavors, I consistently strive to \
-make eco-friendly choices and inspire others to do the same. Embracing \
-the philosophy of sustainable living, I actively support and participate \
-in initiatives that foster a harmonious coexistence with nature, \
-acknowledging our responsibility to protect and preserve the planet for \
-future generations.",
-
-		"Fascinated by the dynamic \
-world of Cloud Computing and DevOps, I am drawn to the transformative \
-power of cloud technologies in modernizing software development and \
-streamlining operations. My passion lies in orchestrating seamless \
-integration, continuous deployment, and automated workflows, harnessing \
-the potential of cloud platforms to create agile and scalable solutions. \
-With a focus on efficiency and reliability, I aspire to unlock the full \
-potential of cloud-native applications, enabling businesses to thrive in \
-the digital era.",
-	];
 
 	const handleCardClick = (index) => {
 		if (index + 1 === selectedCard) {
@@ -114,6 +66,7 @@ the digital era.",
 						key={card.title}
 						icon={card.icon}
 						onClick={handleCardClick}
+						selected={selectedCard === index + 1}
 						{...card}
 					/>
 				))}
