+++ src/constants/index.js
@@ -40,6 +40,61 @@ export const navLinks = [
     },
 ];
 
+
+const titleIntro = [
+    "Overview.",
+    "Backend Developer ☕",
+    "Energy Engineer ⚡",
+    "Green Transition Enthusiast 🌊",
+    "Cloud and Devops ☁️",
+];
+
+const presentationText = [
+    "A passionate Backend Developer and Energy Engineer, currently pursuing \
+an MSc in Software Design at ITU. With a keen interest in exploring the \
+intricacies of new technologies, I immerse myself in the vibrant world \
+of software and hardware development. My aspiration is to uncover the \
+synergy between technology and sustainability, discovering how these two \
+forces can work hand in hand.",
+
+    "As a dedicated Backend Developer, I thrive on \
+crafting the unseen backbone of applications, ensuring seamless data \
+management and optimal performance. My passion lies in designing robust \
+server-side architectures, meticulously handling databases, and building \
+efficient APIs. Embracing the ever-evolving landscape of programming \
+languages and frameworks, I am committed to delivering high-quality, \
+scalable, and secure solutions that empower the frontend to shine.",
+
+    "As an Energy Engineer, I am driven by the goal of \
+promoting sustainable energy practices and reducing the carbon footprint \
+of modern technology. My expertise lies in designing energy-efficient \
+systems, optimizing energy usage, and exploring renewable energy \
+sources. I believe in the transformative power of eco-conscious \
+engineering, and I am dedicated to shaping a greener future through \
+innovation and conscious decision-making.",
+
+    "As a fervent Green Transition Enthusiast, I am deeply committed to \
+advocating for environmental sustainability in every aspect of life. \
+From personal habits to professional endeavors, I consistently strive to \
+make eco-friendly choices and inspire others to do the same. Embracing \
+the philosophy of sustainable living, I actively support and participate \
+in initiatives that foster a harmonious coexistence with nature, \
+acknowledging our responsibility to protect and preserve the planet for \
+future generations.",
+
+    "Fascinated by the dynamic \
+world of Cloud Computing and DevOps, I am drawn to the transformative \
+power of cloud technologies in modernizing software development and \
+streamlining operations. My passion lies in orchestrating seamless \
+integration, continuous deployment, and automated workflows, harnessing \
+the potential of cloud platforms to create agile and scalable solutions. \
+With a focus on efficiency and reliability, I aspire to unlock the full \
+potential of cloud-native applications, enabling businesses to thrive in \
+the digital era.",
+];
+
+
+
 const cardTitles = [
     {
         title: "Backend Developer",
@@ -262,4 +317,4 @@ const projects = [
     },
 ];
 
-export { cardTitles, technologies, experiences, testimonials, projects };
+export { cardTitles, titleIntro, presentationText, technologies, experiences, testimonials, projects };
